[ICC-17] 도메인 package 생성
