+++ src/main/java/com/icc/qasker/quiz/dto/request/ExplanationRequest.java
@@ -1,4 +1,6 @@
-package com.icc.qasker.dto.request;
+package com.icc.qasker.quiz.dto.request;
+
+import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AnswerRequest {
-    private Long problemId;
-    private Long userAnswer;
+public class ExplanationRequest {
+    private List<AnswerRequest> answers;
 }
