+++ src/main/java/com/icc/qasker/quiz/repository/ProblemRepository.java
@@ -1,6 +1,6 @@
-package com.icc.qasker.repository;
+package com.icc.qasker.quiz.repository;
 
-import com.icc.qasker.entity.Problem;
+import com.icc.qasker.quiz.entity.Problem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProblemRepository extends JpaRepository<Problem, Long> {
